IKLOCSEN-1242 Fix light sensor reporting of large lux values
The lux value is supposed to be an unsigned number. In the
HAL is is being assigned to a signed value which causes
large values to incorrectly go negative.

Reviewed-on: http://gerrit.pcs.mot.com/615819
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Sanjay Dhar <sanjaydhar@motorola.com>

+++ stm401_hal.cpp
@@ -483,7 +483,7 @@ int HubSensor::readEvents(sensors_event_t* data, int count)
                 data->version = SENSORS_EVENT_T_SIZE;
                 data->sensor = ID_L;
                 data->type = SENSOR_TYPE_LIGHT;
-                data->light = STM16TOH(buff.data + LIGHT_LIGHT);
+                data->light = (uint16_t)STM16TOH(buff.data + LIGHT_LIGHT);
                 data->timestamp = buff.timestamp;
                 data++;
                 count--;
